// frontend/src/app/openapi/model/project-tool.ts
/*
 * SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
 * SPDX-License-Identifier: Apache-2.0
 *
 * Capella Collaboration
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * Do not edit the class manually.
 + To generate a new version, run `make openapi` in the root directory of this repository.
 */

import { SimpleToolModelWithoutProject } from './simple-tool-model-without-project';
import { SimpleToolVersion } from './simple-tool-version';
import { Tool } from './tool';


export interface ProjectTool { 
    id: number | null;
    tool_version: SimpleToolVersion;
    tool: Tool;
    used_by: Array<SimpleToolModelWithoutProject>;
}