// frontend/src/app/openapi/model/simple-tool-model-without-project.ts
/*
 * SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
 * SPDX-License-Identifier: Apache-2.0
 *
 * Capella Collaboration Manager API
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * Do not edit the class manually.
 + To generate a new version, run `make openapi` in the root directory of this repository.
 */

import { GitModel } from './git-model';


export interface SimpleToolModelWithoutProject { 
    id: number;
    slug: string;
    name: string;
    git_models: Array<GitModel> | null;
}